changed store to search by pounds not credits (29.99)

+++ OOP - Task 1C/Store.h
@@ -9,7 +9,7 @@ class Store
 		Store();
 		~Store();
 		List<Game*> SearchByName(std::string&);
-		List<Game*> SearchByPriceRange(int, int);
+		List<Game*> SearchByPriceRange(double, double);
 		List<Game*> games;
 	private:
 };
