+++ OOP - Task 1C/Utils.h
@@ -17,7 +17,7 @@ public:
     static void recursiveToUpper(char*);
     static std::string toLowerString(std::string); // we don't want to pass by reference as we don't want to altar the base string value
     static std::string toUpperString(std::string);
-    static bool withinPriceRange(int&, int&, int);
+    static bool withinPriceRange(double&, double&, double);
     static int toInt(std::string&);
     static bool SortByName(LibraryItem*, LibraryItem*);
     static bool SortByDate(LibraryItem*, LibraryItem*);
